requirements_agent.txt
neo4j==5.14.1
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
openai==1.6.1
python-dotenv==1.0.0
gradio==4.12.0

